Moved LogRecord to own file
